requirements.txt
fastrtc
fastrtc[vad]
openpyxl
protobuf
python-dotenv 
scipy
sentencepiece
torchaudio
transformers